+++ axelor-project/src/main/java/com/axelor/apps/project/service/TimerTeamTaskServiceImpl.java
@@ -32,7 +32,7 @@ public class TimerTeamTaskServiceImpl extends AbstractTimerService implements Ti
 
     return task.getTimerList()
         .stream()
-        .filter(t -> t.getAssignedTo() == user)
+        .filter(t -> t.getAssignedToUser() == user)
         .findFirst()
         .orElse(null);
   }
@@ -50,7 +50,7 @@ public class TimerTeamTaskServiceImpl extends AbstractTimerService implements Ti
     }
 
     TimerHistory history = new TimerHistory();
-    history.setStartDate(dateTime);
+    history.setStartDateT(dateTime);
     history.setTimer(timer);
     timer.addTimerHistoryListItem(history);